services/audio-voice/requirements.txt
# === CORE WEB FRAMEWORK ===
fastapi==0.120.0
uvicorn[standard]==0.38.0
pydantic==2.12.3
pydantic-settings==2.11.0
python-multipart==0.0.20

# === REDIS & CELERY ===
celery==5.3.4
redis==5.0.1

# === AUDIO PROCESSING ===
torch==2.4.0
torchaudio==2.4.0
numpy==1.26.4
soundfile==0.12.1

# === XTTS (Coqui TTS - ONLY TTS ENGINE) ===
TTS>=0.22.0

# === RVC DEPENDENCIES (Sprint 2) ===
# See docker/requirements-rvc.txt for details
tts-with-rvc==0.1.9
fairseq>=0.12.0
faiss-gpu==1.7.2
librosa==0.10.1
praat-parselmouth==0.4.3
torchcrepe==0.0.23
scipy>=1.11.4
resampy>=0.4.2

# === UTILITIES ===
httpx==0.27.0
